requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langchain>=0.3.7
langchain-groq>=0.2.3
langchain-core>=0.3.29
langsmith>=0.1.147
pydantic==2.10.3
python-dotenv==1.0.1
